fixing bug reading voltages for 2 lines of metadata
The variable 'inds' needs to start at 2 instead of 1. When starting at 1, it reads the wrong cell for voltage, incorrectly calculates the voltage to be 0.008, and then excludes all impacts because it thinks they are below the voltage threshold.

I added .pdf to the gitignore because I want to push this to you now, but the code is plotting temporary PDFs in the background, which are showing up as changes on GitHub.