updates to plotting
CombinePlottedPDFs.m - Previously, this code deleted TransformedData_All.pdf every time it was run and replaced with a new PDF. However, if plotting the events for a single day, this code deleted all previous impact and replaced with only the impacts from the newly processed day instead of appending the new day to the already-plotted events. This update will only delete the "All.pdf" file if all dates are being processed. Otherwise, it will just append impacts to the "All.pdf" file.

PlotData.m - lines 97 and 98 were updated to change the title to X Rot Vel instead of Y Rot vel. Also, added a conditional statement that will only plot the impacts if the PDF doesn't already exist for that day.

TransformData.m - edited line 105 to add all subfolders of the function folder to the path. Subfolders are needed for the plotting of events.